Security hardening: zeroize secrets, SSRF coverage, input validation, docs overhaul
- Add `zeroize` crate for compiler-resistant secret memory clearing (CWE-316)
  with proper Zeroize trait impl and ZeroizeOnDrop for StoredSecret
- Expand scrub_error() to cover Authorization headers, AUTH PLAIN payloads
- Add SSRF protection to SCP (all 4 entry points) and IMAP (list + fetch)
- Add LDAP filter validation (CWE-90: balanced parens, null bytes, length)
- Add GraphQL/Elasticsearch query payload size limits (CWE-400: 1 MiB max)
- Add SNMP walk() weak auth warning (was only on get())
- Add IMAP plaintext connection warnings (CWE-319)
- Add LDAP bind plaintext credential warnings
- Extract security_validate.rs module for protocol-specific validation
- Extract redis_resp.rs module for RESP parser (fix integer cast safety)
- Fix 565 clippy warnings (unnecessary usize casts across all FFI files)
- Enhance security_status() with warnings count, zeroize flag
- Overhaul security docs: architecture, hardening, secrets, DuckDB secrets,
  warnings, SSRF, README, SECURITY.md

+++ src/ffi/grpc.rs
@@ -41,13 +41,13 @@ quack_rs::scalar_callback!(cb_grpc_list_services, |_info, input, output| {
     let mut list_offset: usize = 0;
 
     for row in 0..row_count {
-        let url = unsafe { url_reader.read_str(row as usize) };
+        let url = unsafe { url_reader.read_str(row) };
 
         let result = grpc_reflect::list_services(url);
 
-        unsafe { sw.write_bool(row as usize, 0, result.success) };
+        unsafe { sw.write_bool(row, 0, result.success) };
         unsafe { write_string_list(services_vec, row, &result.services, &mut list_offset) };
-        unsafe { sw.write_varchar(row as usize, 2, &result.message) };
+        unsafe { sw.write_varchar(row, 2, &result.message) };
     }
 });
 
@@ -63,18 +63,18 @@ quack_rs::scalar_callback!(cb_grpc_call, |_info, input, output| {
     let mut sw = unsafe { StructWriter::new(output, 5) };
 
     for row in 0..row_count {
-        let url = unsafe { url_reader.read_str(row as usize) };
-        let service = unsafe { svc_reader.read_str(row as usize) };
-        let method = unsafe { method_reader.read_str(row as usize) };
-        let payload = unsafe { payload_reader.read_str(row as usize) };
+        let url = unsafe { url_reader.read_str(row) };
+        let service = unsafe { svc_reader.read_str(row) };
+        let method = unsafe { method_reader.read_str(row) };
+        let payload = unsafe { payload_reader.read_str(row) };
 
         let result = grpc::call(url, service, method, payload);
 
-        unsafe { sw.write_bool(row as usize, 0, result.success) };
-        unsafe { sw.write_i32(row as usize, 1, result.status_code) };
-        unsafe { sw.write_varchar(row as usize, 2, &result.body) };
-        unsafe { sw.write_i32(row as usize, 3, result.grpc_status) };
-        unsafe { sw.write_varchar(row as usize, 4, &result.grpc_message) };
+        unsafe { sw.write_bool(row, 0, result.success) };
+        unsafe { sw.write_i32(row, 1, result.status_code) };
+        unsafe { sw.write_varchar(row, 2, &result.body) };
+        unsafe { sw.write_i32(row, 3, result.grpc_status) };
+        unsafe { sw.write_varchar(row, 4, &result.grpc_message) };
     }
 });
